requirements.txt
fbs 
PyQt5
PyQt6
numpy
# numpy; python_version >= '3', installed in coin_*_instructions.py atm
# see https://tinyurl.com/y3dm3h86 for details
# setuptools from v45+ removed the support for Python 2, so we pin an old release
setuptools
setuptools
sphinx
six
wheel
PyOpenGL
PySide2
mysql-connector-python 
Selenium
webdriver_manager
auto-py-to-exe
